tests/requirements.txt
pytest>=7.0
python-hcl2>=4.0
